Change the threadsafe_queue to store shared_ptr<T> instead of T
The pop functions that rake a reference to a variable to receive the new
value now have to reference tthe stored pointer, but the pop functions
that return a std::shared_ptr<T> instance can retrieve it from the
queue.

There is an additional benefit: the allocation of the new instance can
now be done outside the lock in push(), while in the previous version it
had to be done while holding the lock in pop(). Because memory
allocation is typically quite an expensive operation, this can be
beneficial for the performance of the queue, because it reduces the time
the mutex is held, allowing other threads to perform operation in the
queue in the meantime.